[lib] Move match_path() and ignore_path() to paths.c
The match_path() function was static to ignore.c, but make it public
and move it to paths.c.  Also move ignore_path() to paths.c since it
was the only function remaining in ignore.c.
